First Kivy Project
fun thing I did, hopefully someone can get use out of this

sorry for not using many comments, I was not expecting to share this with others. Please let me know how i could improve the code or if something does not work, or if there is a better way to do it. Im open to listening all the time.